Rename msb to find_last_set and make it 32 bit

+++ include/sfc.h
@@ -9,6 +9,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 #define SFC_HEADER_TITLE_MAX_SIZE 21
 
@@ -58,10 +59,10 @@ bool sfc_header_set_map(sfc_header header, enum sfc_map map);
 bool sfc_header_chipset(sfc_header header, struct sfc_chipset * chipset);
 bool sfc_header_set_chipset(sfc_header header, struct sfc_chipset chipset);
 
-size_t sfc_header_rom_size(sfc_header header);
-bool sfc_header_set_rom_size(sfc_header header, size_t size);
+uint_least32_t sfc_header_rom_size(sfc_header header);
+bool sfc_header_set_rom_size(sfc_header header, uint_least32_t size);
 
-size_t sfc_header_ram_size(sfc_header header);
-bool sfc_header_set_ram_size(sfc_header header, size_t size);
+uint_least32_t sfc_header_ram_size(sfc_header header);
+bool sfc_header_set_ram_size(sfc_header header, uint_least32_t size);
 
 #endif
